fix for bug 1948533
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2957 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/test/CuTest.cpp
@@ -377,7 +377,7 @@ void CuSuiteSummary(CuSuite* testSuite, CuString* summary, bool times)
 		int bufferLen = 25-summary->length-10;
 		for (int i=0;i<bufferLen;i++ )
 			CuStringAppend(summary,_T(" "));
-		CuStringAppendFormat(summary,_T(" - %dms"),testSuite->timeTaken);
+		CuStringAppendFormat(summary,_T(" - %dms"), (int32_t)testSuite->timeTaken);
 	}
 	CuStringAppend(summary, _T("\n"));
 }
